[1639] Ядро / Заклинания: исправлен эксплойт, позволяющий использовать предметы, несмотря на общее время восстановления. Так же мелкие правки к предыдущим коммитам.

+++ src/server/game/Entities/Unit/Unit.cpp
@@ -11987,7 +11987,7 @@ uint32 Unit::SpellDamageBonusTaken(Unit* caster, SpellInfo const* spellProto, ui
         AddPct(TakenTotalMod, mult);
         if (GetTypeId() == TYPEID_UNIT)
         {
-            float u_mult = GetTotalAuraModifierByMiscMask(SPELL_AURA_MOD_CREATURE_AOE_DAMAGE_AVOIDANCE, spellProto->SchoolMask);
+            int32 u_mult = GetTotalAuraModifierByMiscMask(SPELL_AURA_MOD_CREATURE_AOE_DAMAGE_AVOIDANCE, spellProto->SchoolMask);
             AddPct(TakenTotalMod, u_mult);
         }
     }
